Update from github
testing to see when you edit a code file in github and commit its changes you can see it changed her in github one time. but not replit. you have to go to replit and then synchronize w the remote branch of the repo. press refresh and fetch if not coming up yet